+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.15.0 (2025-12-13)
+
+- Add support for setting custom server endpoint and endpoint URL
+- Add chat/completions endpoint support for custom server for compatibility with servers such as Google Gemini
+
 ## 2.14.0 (2025-12-13)
 
 - Add GPT 5.2 and GPT 5.2 Pro chat models